Kompilierung gefixt nach den letzten übernommenen Commits.

+++ src/server/game/Server/Protocol/Handlers/MovementHandler.cpp
@@ -379,7 +379,7 @@ void WorldSession::HandleMovementOpcodes(WorldPacket & recv_data)
     {
         plMover->UpdateFallInformationIfNeed(movementInfo, opcode);
 
-        uint32 underMapValue = -500.0f;
+        float underMapValue = -500.0f;
         switch (plMover->GetMapId())
         {
             case 617: underMapValue = 3.0f;  break; // Dalaran Sewers
